feat: implement local hoogle search
