+++ README.md
@@ -24,6 +24,7 @@ Supercharge your Haskell experience in [neovim](https://neovim.io/)!
 ### Optional
 
 * [`telescope.nvim`](https://github.com/nvim-telescope/telescope.nvim)
+* A local [`hoogle`](https://github.com/ndmitchell/hoogle/blob/master/docs/Install.md) installation
 
 ## Installation
 
@@ -54,20 +55,23 @@ To get started quickly with the default setup, add the following to your NeoVim 
 ```lua
 local ht = require('haskell-tools')
 
+local opts = { noremap = true, silent = true, buffer = bufnr }
 ht.setup {
   hls = {
     -- See nvim-lspconfig's  suggested configuration for keymaps, etc.
     on_attach = function(client, bufnr)
-      local opts = { noremap = true, silent = true, buffer = bufnr }
       -- haskell-language-server relies heavily on codeLenses,
       -- so auto-refresh (see advanced configuration) is enabled by default
       vim.keymap.set('n', '<space>ca', vim.lsp.codelens.run, opts)
-      vim.keymap.set('n', '<space>hs', ht.hoogle_signature, opts)
+      vim.keymap.set('n', '<space>hs', ht.hoogle.hoogle_signature, opts)
       -- default_on_attach(client, bufnr)  -- if defined, see nvim-lspconfig
     end,
   },
 }
 ```
+
+If using a local `hoogle` installation, [follow these instructions](https://github.com/ndmitchell/hoogle/blob/master/docs/Install.md#generate-a-hoogle-database)
+to generate a database.
 
 ## Features
 
@@ -82,28 +86,35 @@ ht.setup {
 - [x] Automatically refreshes code lenses by default, which haskell-language-server heavily relies on. [Can be disabled.](#advanced-configuration)
 - [x] The following code lenses are currently supported:
 
-##### [Show/Add type signatures for bindings without type signatures](https://haskell-language-server.readthedocs.io/en/latest/features.html#add-type-signature)
+#### [Show/Add type signatures for bindings without type signatures](https://haskell-language-server.readthedocs.io/en/latest/features.html#add-type-signature)
 [![asciicast](https://asciinema.org/a/zC88fqMhPq25lHFYgEF6OxMgk.svg)](https://asciinema.org/a/zC88fqMhPq25lHFYgEF6OxMgk?t=0:04)
 
-##### [Evaluate code snippets in comments](https://haskell-language-server.readthedocs.io/en/latest/features.html#evaluation-code-snippets-in-comments)
+#### [Evaluate code snippets in comments](https://haskell-language-server.readthedocs.io/en/latest/features.html#evaluation-code-snippets-in-comments)
 [![asciicast](https://asciinema.org/a/TffryPrWpBkLnBK6dKXvOxd41.svg)](https://asciinema.org/a/TffryPrWpBkLnBK6dKXvOxd41?t=0:04)
 
-##### [Make import lists fully explicit](https://haskell-language-server.readthedocs.io/en/latest/features.html#make-import-lists-fully-explicit-code-lens)
+#### [Make import lists fully explicit](https://haskell-language-server.readthedocs.io/en/latest/features.html#make-import-lists-fully-explicit-code-lens)
 [![asciicast](https://asciinema.org/a/l2ggVaN5eQbOj9iGkaethnS7P.svg)](https://asciinema.org/a/l2ggVaN5eQbOj9iGkaethnS7P?t=0:02)
 
-##### [Fix module names that do not match the file path](https://haskell-language-server.readthedocs.io/en/latest/features.html#fix-module-names)
+#### [Fix module names that do not match the file path](https://haskell-language-server.readthedocs.io/en/latest/features.html#fix-module-names)
 [![asciicast](https://asciinema.org/a/n2qd2zswLOonl2ZEb8uL4MHsG.svg)](https://asciinema.org/a/n2qd2zswLOonl2ZEb8uL4MHsG?t=0:02)
 
 ### Beyond haskell-language-server
 
 The below features are not implemented by haskell-language-server.
 
-##### Hoogle-search for signature
-Search for the type signature under the cursor.
-Falls back to the word under the cursor if the type signature cannot be determined.
+#### Hoogle-search for signature
+
+* Search for the type signature under the cursor.
+* Falls back to the word under the cursor if the type signature cannot be determined.
+* Telescope keymaps:
+  - `<CR>` to copy the selected entry to the clipboard.
+  - `<C-b>` to open the selected entry's URL in a browser.
+
 ```lua
-require('haskell-tools').hoogle_signature()
+require('haskell-tools').hoogle.hoogle_signature()
 ```
+
+[![asciicast](https://asciinema.org/a/4GSmXrCvpt7idBHnuZVQQkJ9R.svg)](https://asciinema.org/a/4GSmXrCvpt7idBHnuZVQQkJ9R)
 
 ### Planned
 
@@ -115,18 +126,26 @@ For planned features, refer to the [issues](https://github.com/MrcJkb/haskell-to
 To modify the language server configs, call
 
 ```lua
+-- defaults
 require('haskell-tools').setup {
   tools = { -- haskell-tools options
     codeLens = {
       -- Whether to automatically display/refresh codeLenses
-      autoRefresh = false, -- defaults to true
+      autoRefresh = true,
+    },
+    hoogle = {
+      -- 'auto': Choose a mode automatically, based on what is available.
+      -- 'telescope-local': Force use of a local installation.
+      -- 'telescope-web': The online version (depends on curl).
+      -- 'browser': Open hoogle search in the default browser.
+      mode = 'auto', 
     },
   },
   hls = { -- LSP client options
     -- ...
     haskell = { -- haskell-language-server options
-      formattingProvider = 'fourmolu', -- Defaults to 'ormolu'
-      checkProject = false, -- Defaults to true, which could have a performance impact on large monorepos.
+      formattingProvider = 'ormolu', 
+      checkProject = true, -- Setting this to true could have a performance impact on large monorepos.
       -- ...
     }
   }
